// apps/portal-api/src/model/kanel/public/ActionTracking.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

/** Identifier type for public.ActionTracking */
export type ActionTrackingId = string & { __brand: 'public.ActionTracking' };

/** Represents the table public.ActionTracking */
export default interface ActionTracking {
  id: ActionTrackingId;

  contextual_id: string;

  status: string | null;

  created_at: Date;

  ended_at: Date | null;

  type: string;
}

/** Represents the initializer for the table public.ActionTracking */
export interface ActionTrackingInitializer {
  id: ActionTrackingId;

  contextual_id: string;

  /** Default value: 'INITIALISE'::character varying */
  status?: string | null;

  /** Default value: CURRENT_TIMESTAMP */
  created_at?: Date;

  ended_at?: Date | null;

  type: string;
}

/** Represents the mutator for the table public.ActionTracking */
export interface ActionTrackingMutator {
  id?: ActionTrackingId;

  contextual_id?: string;

  status?: string | null;

  created_at?: Date;

  ended_at?: Date | null;

  type?: string;
}